requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0

onnx>=1.14.0
onnxruntime>=1.15.0

# Web application
gradio>=4.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Image processing
pillow>=10.0.0

# Numerical computing
numpy>=1.24.0

# Jupyter for experiments
jupyter>=1.0.0
ipykernel>=6.25.0